rename: dispatchData to actionType
